requirements.txt
adafruit-circuitpython-dht~=3.7.4
board~=1.0
paho-mqtt~=1.6.1
guizero~=1.4.0